+++ mathics/builtin/drawing/graphics3d.py
@@ -77,6 +77,10 @@ class Graphics3D(Graphics):
         >> Graphics3D[Polygon[{{0,0,0}, {0,1,1}, {1,0,0}}]]
          = -Graphics3D-
 
+        The 'Background' option allows to set the color of the background:
+        >> Graphics3D[Sphere[], Background->RGBColor[.6, .7, 1.]]
+         = -Graphics3D-
+
         In 'TeXForm', 'Graphics3D' creates Asymptote figures:
         >> Graphics3D[Sphere[]] // TeXForm
          = #<--#